Refactor stick-header into a module
